+++ gradience/backend/models/repo.py
@@ -16,7 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-from gradience.backend.utils.utils import to_slug_case
+from gradience.backend.utils.common import to_slug_case
 from gradience.backend.models.preset import Preset, presets_dir
 import os
