+++ gradience/backend/utils/common.py
@@ -1,4 +1,4 @@
-# utils.py
+# common.py
 #
 # Change the look of Adwaita, with ease
 # Copyright (C) 2022 Gradience Team
