+++ gradience/backend/utils/meson.build
@@ -2,6 +2,6 @@ utilsdir = 'gradience/backend/utils'
 
 gradience_sources = [
     '__init__.py',
-    'utils.py'
+    'common.py'
 ]
 PY_INSTALLDIR.install_sources(gradience_sources, subdir: utilsdir)
