+++ gradience/frontend/widgets/explore_preset_row.py
@@ -20,7 +20,7 @@ import os
 
 from gi.repository import Gtk, Adw
 
-from gradience.backend.utils.utils import to_slug_case, buglog
+from gradience.backend.utils.common import to_slug_case, buglog
 from gradience.backend.preset_downloader import download_preset
 from gradience.backend.constants import rootdir
